Moved creation of enum type to test scope provider.
