+++ test-plugins/org.yakindu.sct.simulation.core.sexec.test/src/org/yakindu/sct/model/sexec/interpreter/test/STextInterpreterTest.java
@@ -12,9 +12,6 @@ package org.yakindu.sct.model.sexec.interpreter.test;
 
 import static org.junit.Assert.assertEquals;
 
-import org.eclipse.emf.common.util.URI;
-import org.eclipse.emf.ecore.resource.Resource;
-import org.eclipse.emf.ecore.resource.impl.ResourceImpl;
 import org.eclipse.xtext.junit4.InjectWith;
 import org.eclipse.xtext.junit4.XtextRunner;
 import org.junit.After;
@@ -22,9 +19,6 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.yakindu.base.expressions.expressions.Expression;
-import org.yakindu.base.types.EnumerationType;
-import org.yakindu.base.types.Enumerator;
-import org.yakindu.base.types.TypesFactory;
 import org.yakindu.base.types.typesystem.GenericTypeSystem;
 import org.yakindu.base.types.typesystem.ITypeSystem;
 import org.yakindu.sct.model.sgraph.Scope;
@@ -479,7 +473,7 @@ public class STextInterpreterTest extends AbstractSTextTest {
 	@Test
 	public void testEnumAssignment() {
 		execute("internal: var enumVar : EnumType", "enumVar = EnumType.B");
-		assertEquals(new Long(((EnumerationType)typeSystem.getType("EnumType")).getEnumerator().get(1).getLiteralValue()), getEnumValue());
+		assertEquals(1L, getEnumValue());
 	}
 	
 	@Test
@@ -551,9 +545,7 @@ public class STextInterpreterTest extends AbstractSTextTest {
 		ExecutionVariable enumVar = new ExecutionVariableImpl();
 		enumVar.setName("enumVar");
 		enumVar.setFqName("enumVar");
-		EnumerationType enumType = createEnumType();
-		enumVar.setType(enumType);
-		enumVar.setValue(enumType.getEnumerator().get(0));
+		enumVar.setType(typeSystem.getType(GenericTypeSystem.INTEGER));
 		context.getSlots().add(enumVar);
 		
 		CompositeSlot cpVar = new CompositeSlotImpl();
@@ -567,23 +559,6 @@ public class STextInterpreterTest extends AbstractSTextTest {
 		featureVar.setValue(0);
 		cpVar.getSlots().add(featureVar);
 		context.getSlots().add(cpVar);
-	}
-	
-	private EnumerationType createEnumType() {
-		EnumerationType enumType = TypesFactory.eINSTANCE.createEnumerationType();
-		enumType.setName("EnumType");
-		
-		Enumerator enumA = TypesFactory.eINSTANCE.createEnumerator();
-		enumA.setName("A");
-		enumType.getEnumerator().add(enumA);
-		
-		Enumerator enumB = TypesFactory.eINSTANCE.createEnumerator();
-		enumB.setName("B");
-		enumType.getEnumerator().add(enumB);
-		
-		Resource resource = new ResourceImpl(URI.createURI("types2"));
-		resource.getContents().add(enumType);
-		return enumType;
 	}
 
 	protected Object getBoolValue() {
